domd: Wrap VIS related code by AOS_VIS_PLUGINS
We should perform VIS related actions only if corresponding
option is enabled. This means that only if AOS_VIS_PLUGINS
is not empty then we need to add meta-aos layer and add
IMAGE_INSTALL targets.
And if option is empty we need to mask aos_vis related bbappend
to avoid warning "No recipes available for".

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babcuk@epam.com>